requirements.txt
transformers==4.39.3
streamlit==1.33.0
tensorflow==2.15.0
pdf2image==1.17.0
pypdf2==3.0.1
langchain-groq==0.1.0
langchain
# ragatouille
faiss-gpu
langchain-core==0.1.40
langchain-community==0.0.31
llama-parse
python-dotenv==1.0.1
langchain-text-splitters
langchain-voyageai
langchain-experimental